requirements.txt
torchvision==0.15.1